+++ src/main/java/uz/pdp/config/S3Config.java
@@ -1,15 +1,15 @@
 package uz.pdp.config;
 
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
 import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.regions.Regions;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
 
 @Configuration
 public class S3Config {
@@ -21,6 +21,7 @@ public class S3Config {
 
     @Value("${aws.s3.bucket}")
     private String bucketName;
+
 
     @Bean
     public AmazonS3 s3Client() {