+++ src/main/java/uz/pdp/controller/AdminController.java
@@ -3,6 +3,7 @@ package uz.pdp.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
+import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,12 +13,14 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
 import uz.pdp.dto.SellerRequestDto;
 import uz.pdp.payload.EntityResponse;
 import uz.pdp.service.AdminService;
 
 @RestController
 @RequestMapping("/api/admin")
+@Slf4j
 @Tag(name = "Admin Control", description = "APIs for admin control and management")
 public class AdminController {
     private static final Logger logger = LoggerFactory.getLogger(AdminController.class);