+++ src/main/java/uz/pdp/dto/VerificationRequest.java
@@ -1,5 +1,6 @@
 package uz.pdp.dto;
 
+import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -8,8 +9,10 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class SellerRequestDto {
+public class VerificationRequest {
     @NotNull(message = "User ID is required")
     private Long userId;
     
+    @NotBlank(message = "Verification code is required")
+    private String code;
 }